[models] Added swappable models #10
Removed base admin
Moved all handlers from models.py > handlers.py
Added sample app
