+++ .travis.yml
@@ -34,6 +34,7 @@ before-script:
   - ./run-qa-checks
 
 script:
+  - SAMPLE_APP=1 coverage run --source=openwisp_notifications runtests.py
   - coverage run -a --source=openwisp_notifications runtests.py
 
 after_success:
