+++ openwisp_notifications/apps.py
@@ -4,6 +4,7 @@ from django.utils.translation import ugettext_lazy as _
 
 class OpenwispNotificationsConfig(AppConfig):
     name = 'openwisp_notifications'
+    label = 'openwisp_notifications'
     verbose_name = _('Notifications')
 
     def ready(self):
