+++ openwisp_notifications/templatetags/notification_tags.py
@@ -1,11 +1,13 @@
 # -*- coding: utf-8 -*-
-import swapper
 from django.core.cache import cache
 from django.template import Library
 from django.utils.html import format_html
-from notifications.templatetags.notifications_tags import notifications_unread
+from notifications.templatetags.notifications_tags import (
+    notifications_unread as base_notification_unread,
+)
+from openwisp_notifications.swapper import load_model
 
-Notification = swapper.load_model('openwisp_notifications', 'Notification')
+Notification = load_model('Notification')
 
 register = Library()
 
@@ -15,7 +17,7 @@ def get_notifications_count(context):
     cache_key = Notification.COUNT_CACHE_KEY.format(user_pk)
     count = cache.get(cache_key)
     if count is None:
-        count = notifications_unread(context)
+        count = base_notification_unread(context)
         count = '99+' if count > 99 else count
         cache.set(cache_key, count)
     return count
