+++ requirements.txt
@@ -2,5 +2,5 @@ django-notifications-hq>=1.6,<1.7
 openwisp-users>=0.2.0,<0.3.0
 # TODO: enable when openwisp-utils 0.5 is released
 # openwisp-utils[rest]~=0.5
-swapper
+swapper~=1.1
 markdown>=3.2,<3.3
