+++ tests/openwisp2/celery.py
