+++ tests/openwisp2/sample_notifications/admin.py
@@ -1,0 +1,4 @@
+from openwisp_notifications.admin import (  # noqa
+    NotificationAdmin,
+    NotificationUserInline,
+)
