+++ tests/openwisp2/sample_notifications/templatetags/notification_tags.py
@@ -1,0 +1,1 @@
+from openwisp_notifications.templatetags.notification_tags import register  # noqa
